requirements.txt
# Core Django and openIMIS dependencies
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-extensions>=3.2.0

# AI and Machine Learning dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.10.0

# Deep Learning (optional for advanced features)
tensorflow>=2.13.0
torch>=2.0.0

# Data visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Natural Language Processing
nltk>=3.8.0
spacy>=3.6.0
transformers>=4.30.0

# API and web services
requests>=2.31.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Database and caching
redis>=4.6.0
celery>=5.3.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing and development
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.4.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
click>=8.1.0
tqdm>=4.65.0 